make detect empty collectors more reliable
im an autohotkey enthusiast and i have been making bots for 2 years.
i made a script for clash of clans last year. for auto search and auto make troops,
then after that nothing else.. when i saw your program i was amazed. i have tried many other bots out there. but this is the most efficient.
one thing that i dont like about the bot is that. sometimes... (30% of the time) it will attack a base because it has a lvl 12 collectors and it is half full, so i have been reading the program lately to see. the detect empty collectors is almost flawless..
2 things i think will make it better. 
(btw im not a java programmer so... bare with me)
this part here i think need to add something
	if (ConfigUtils.instance().doConditionsMatch(gold, elixir, de) &&
				(!ConfigUtils.instance().isDetectEmptyCollectors() || ImageParser.isCollectorFullBase())) {
something like.
a new string that will detect the tiny gold collectors? because the elixir collectors are not very visual reliable. we both know that.
and i know you did not make it gold collectors is that the picture of that tiny box at the gold collectors are hidden.. but hey, you don't need to see 6 boxes... you only need to see 1 to know if all 6 has something inside it.
secondly, to make the detect empty collectors more reliable.. detect gold storages pictures? if the bot sees the storage with coins press next?
thats all ^_^ i dont force you with anything but theres a proposal change here. and i cant find the forums 

if empty collectors is true
if gold collectors is true
if somethinginside storage is true
stateattack